Rename and re-namespace Assertion/ExpectedValue
