+++ src/Assertion/ComparisonAssertionInterface.php
@@ -2,10 +2,10 @@
 
 namespace webignition\BasilModel\Assertion;
 
-use webignition\BasilModel\Value\Assertion\ExpectedValueInterface;
+use webignition\BasilModel\Value\ExpectableValueInterface;
 
 interface ComparisonAssertionInterface extends ExaminationAssertionInterface
 {
-    public function getExpectedValue(): ExpectedValueInterface;
-    public function withExpectedValue(ExpectedValueInterface $value): ComparisonAssertionInterface;
+    public function getExpectedValue(): ExpectableValueInterface;
+    public function withExpectedValue(ExpectableValueInterface $value): ComparisonAssertionInterface;
 }
